Move FolderProcessorSettings to the models project
