+++ src/FolderProcessor.Extensions.Microsoft.DependencyInjection/FolderProcessor.Extensions.Microsoft.DependencyInjection.csproj
@@ -27,4 +27,8 @@
       <ProjectReference Include="..\FolderProcessor\FolderProcessor.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Models" />
+    </ItemGroup>
+
 </Project>
